fix: added missing navigation bar
