+++ app/src/androidTest/java/com/github/se/assocify/screens/TreasuryScreenTest.kt
@@ -6,10 +6,12 @@ import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.github.se.assocify.navigation.NavigationActions
+import com.github.se.assocify.ui.screens.TreasuryScreen
 import com.github.se.assocify.ui.screens.treasury.TreasuryScreen
 import com.kaspersky.components.composesupport.config.withComposeSupport
 import com.kaspersky.kaspresso.kaspresso.Kaspresso
 import com.kaspersky.kaspresso.testcases.api.testcase.TestCase
+import io.github.kakaocup.compose.node.element.ComposeScreen
 import io.mockk.every
 import io.mockk.mockk
 import org.junit.Before
@@ -40,6 +42,36 @@ class TreasuryScreenTest : TestCase(kaspressoBuilder = Kaspresso.Builder.withCom
     with(composeTestRule) {
       onNodeWithTag("mainNavBarItem/home").performClick()
       assert(tabSelected)
+    }
+  }
+
+  @Test
+  fun testTabSwitching() {
+    ComposeScreen.onComposeScreen<TreasuryScreen>(composeTestRule) {
+      budgetTab.assertIsDisplayed()
+      budgetTab.performClick()
+      budgetTab.assertIsSelected()
+
+      balanceTab.assertIsDisplayed()
+      balanceTab.performClick()
+      balanceTab.assertIsSelected()
+
+      myReceiptsTab.assertIsDisplayed()
+      myReceiptsTab.performClick()
+      myReceiptsTab.assertIsSelected()
+    }
+  }
+
+  @Test
+  fun testTodoListItem() = run {
+    ComposeScreen.onComposeScreen<TreasuryScreen>(composeTestRule) { receiptItemBox.assertIsDisplayed() }
+  }
+
+  @Test
+  fun createTodo() = run {
+    ComposeScreen.onComposeScreen<TreasuryScreen>(composeTestRule) {
+      createReceiptFab.assertIsDisplayed()
+      createReceiptFab.performClick()
     }
   }
 }
